Modified to use indices that are shorts, versus integers.  This patch is brought to you because Android just wants to be that guy...
Former-commit-id: 8f923f4d44ffd7c1fa3820f1d826bf27d2447a99

+++ Sample/source/Mesh.cpp
@@ -138,7 +138,7 @@ std::vector<GLfloat>& Mesh::Data()
     return data;
 }
 
-std::vector<GLuint>& Mesh::Indices()
+std::vector<GLushort>& Mesh::Indices()
 {
     return indices;
 }
